% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTarget.R
\name{addTarget}
\alias{addTarget}
\title{Add a new target to TAS without re-running taShrink()}
\usage{
addTarget(X, TASoutput, NEWtarget)
}
\arguments{
\item{X}{\code{matrix} -- data matrix with variables in rows and 
observations in columns. This method performs best when there are more 
variables than observations.}

\item{TASoutput}{\code{list} -- output from the taShrink function.}

\item{NEWtarget}{\code{matrix} -- a new target to add to the shrinkage
estimation. Must have the same dimensions as the other targets.}
}
\value{
\code{list} -- the updated TAS output having added the new target matrix
 to the target set.
}
\description{
Add a new target to TAS without re-running taShrink()
}
