% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTarget}
\alias{getTarget}
\title{Construct a target matrix for single target linear shrinkage}
\usage{
getTarget(X, varNumber = 2L, corNumber = 1L)
}
\arguments{
\item{X}{\code{matrix} -- data matrix with variables in rows and 
observations in columns.}

\item{varNumber}{\code{numeric} -- c(1, 2, 3) variance structure for the
 target matrix.
1 sets all variances equal to 1. 2 sets all variances equal to their sample
 mean (using \code{X}).
3. sets all variances to their sample values (using \code{X}).}

\item{corNumber}{\code{numeric} -- c(1, 2, 3) correlation structure for
 the target matrix. 1 sets the correlations to 0. 2 sets the correlations
  equal 
to their sample mean (using \code{X}). 3 sets the correlations equals to an
 autocorrelation
structure with parameter equal to the sample mean (using \code{X}).}
}
\value{
\code{matrix} -- target matrix for linear shrinkage
 estimation.
}
\description{
Construct a popular target matrix from the linear shrinkage literature.
These targets consist of a combination of variance and correlation
 structure. Possible variance structures are unit, sample mean, and 
 sample. Possible correlation structures are zero, sample mean, and
  autocorrelation.
}
