% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTargetSet}
\alias{getTargetSet}
\title{Construct a set of target matrices for Target-Averaged linear shrinkage}
\usage{
getTargetSet(X)
}
\arguments{
\item{X}{\code{matrix} -- data matrix with variables in rows and 
observations in columns.}
}
\value{
\code{array} -- a \code{p}x\code{p}x9 array of
target matrices, where \code{p} is the number of variables of \code{X}.
}
\description{
Construct a set of popular target matrices from the linear shrinkage
 literature.
These nine targets consist of the combinations of variance and correlation
 structures; variance structures are unit, sample mean, and 
 sample; correlation structures are zero, sample mean, and
autocorrelation.
}
\examples{
  set.seed(102)
  X <- matrix(rnorm(50), 10, 5) # p=10, n=5, identity covariance
  X <- t(scale(t(X), center=TRUE, scale=FALSE)) # mean 0
  ts <- getTargetSet(X) # an array of targets
  # inspect the variances of the targets
  vars <- apply(ts, 3, diag)
  colnames(vars) <- paste("target", c(1:9), sep="")
  vars
  boxplot(vars, ylab = "variances")
  # inspect the correlations of the targets
  corrs <- apply(ts, 3, function(x){cov2cor(x)[lower.tri(x)]})
  colnames(corrs) <- paste("target", c(1:9), sep="")
  corrs
  boxplot(corrs, ylab = "correlations")
}
\seealso{
\code{\link{taShrink}}
}
