% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTargetSet}
\alias{getTargetSet}
\title{Construct a set of target matrices for Target-Averaged linear shrinkage}
\usage{
getTargetSet(X)
}
\arguments{
\item{X}{\code{matrix} -- data matrix with variables in rows and 
observations in columns.}
}
\value{
\code{array} -- a \code{p}x\code{p}x9 array of
target matrices, where \code{p} is the number of variables of \code{X}.
}
\description{
Construct a set of popular target matrices from the linear shrinkage
 literature.
These nine targets consist of the combinations of variance and correlation
 structures; variance structures are unit, sample mean, and 
 sample; correlation structures are zero, sample mean, and
autocorrelation.
}
