% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetWeights.R
\name{targetWeights}
\alias{targetWeights}
\title{Extract the target-specific shrinkage weights from TAS output}
\usage{
targetWeights(TASoutput)
}
\arguments{
\item{TASoutput}{\code{list} -- output from the taShrink function.}
}
\value{
\code{list} -- the weights from each target matrix and the sample
covariance matrix in the target-averaged linear shrinkage.
}
\description{
Extract the target-specific shrinkage weights from TAS output
}
