\name{TAS-package}
\alias{TAS-package}
\alias{TAS}
\docType{package}
\title{
  Target-Averaged linear Shrinkage estimation
}
\description{
  Conjugate Bayesian covariance matrix estimation
  using linear shrinkage with multiple target matrices. Most       useful in high-dimensional data settings, where the number of    variables (p) is greater than the number of samples (n).
}
\details{
  This package contains functions for 
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Harry Gray

Maintainer: Harry Gray <harry.gray@mrc-bsu.cam.ac.uk>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
